requirements.txt
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.19
PyYAML==3.11
smartypants==1.8.3
typogrify==2.0.4
wsgiref==0.1.2
